First details display

+++ app/src/main/java/com/albinpoignot/foursquarepoc/models/Venue.java
@@ -13,7 +13,7 @@ public class Venue extends CompactVenue
 
     private Integer price;
 
-    private Integer rating;
+    private Float rating;
 
     public String getDescription()
     {
@@ -55,12 +55,12 @@ public class Venue extends CompactVenue
         this.price = price;
     }
 
-    public Integer getRating()
+    public Float getRating()
     {
         return rating;
     }
 
-    public void setRating(Integer rating)
+    public void setRating(Float rating)
     {
         this.rating = rating;
     }
